+++ Dyalect/Compiler/OpSizeHelper.cs
@@ -32,6 +32,7 @@
             0, //Div
             0, //Rem
             0, //Neg
+            0, //Plus
             0, //Not
             0, //BitNot
             0, //Len