+++ Dyalect/Compiler/OpStackHelper.cs
@@ -32,6 +32,7 @@
             -1, //Div
             -1, //Rem
             0,  //Neg
+            0,  //Plus
             0,  //Not
             0,  //BitNot
             0,  //Len