+++ Dyalect/Parser/Model/UnaryOperator.cs
@@ -5,6 +5,7 @@
         None,
         Not,
         Neg,
+        Plus,
         BitwiseNot,
         Length
     }